reorganise content
